comps/embeddings/multimodal_clip/requirements.txt
dateparser
docarray[full]
einops
fastapi
huggingface_hub
langchain
open_clip_torch
opentelemetry-api
opentelemetry-exporter-otlp
opentelemetry-sdk
prometheus-fastapi-instrumentator
sentence_transformers
shortuuid
uvicorn
